--- Constants.java
@@ -1,39 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot;
-
-/**
- * The Constants class provides a convenient place for teams to hold robot-wide numerical or boolean
- * constants. This class should not be used for any other purpose. All constants should be declared
- * globally (i.e. public static). Do not put anything functional in this class.
- *
- * <p>It is advised to statically import this class (or one of its inner classes) wherever the
- * constants are needed, to reduce verbosity.
- */
-public final class Constants {
-   public static final class RobotMap{
-        
-    
-        //these ID's are accurate to the old robot, not the 2022 robit
-
-        public static final int MOTOR_LEFT_MASTER_ID = 1;
-        public static final int MOTOR_LEFT_SLAVE_ID = 2;
-        public static final int MOTOR_RIGHT_MASTER_ID = 4;
-        public static final int MOTOR_RIGHT_SLAVE_ID = 3;
-        // joystick port, may changes if there is a mouse or something plugged in, be careful about that
-        public static final int JOYSTICK_PORT = 0;
-
-
-
-
-
-   } 
-
-
-
-
-
-
-}
